Refactor to be based on the Document base class
